refactor: inference code refactoring
